requirements.txt
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
mutagen==1.47.0
packaging==23.2
pillow==10.2.0
pycryptodomex==3.20.0
requests==2.31.0
tksheet==7.0.5
urllib3==2.2.1
websockets==12.0
yt-dlp==2023.12.30
